+++ c_dal/c_dal.csproj
@@ -42,7 +42,14 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Acceso_Datos.cs" />
+    <Compile Include="CRUD.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\c_entidades\c_entidades.csproj">
+      <Project>{502AC588-8E04-48E5-A710-0CEFC7F0CDEB}</Project>
+      <Name>c_entidades</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
